A couple more changes to my delete method
I'm not sure if you're going to test it, but I realized I forgot to decrement the totalNodes in the delete method. I also created an out so that if the key we're trying to delete doesn't exist, it will exit. I think there could be some null pointer exceptions elsewise.